Move shared util to `kubeojects/capability`
