+++ src/kubeobjects/capability.go
@@ -1,10 +1,8 @@
-package testinghelpers
+package kubeobjects
 
 import dynatracev1beta1 "github.com/Dynatrace/dynatrace-operator/src/api/v1beta1"
 
-type DtCapability = dynatracev1beta1.CapabilityDisplayName
-
-func DoTestSetCapability(instance *dynatracev1beta1.DynaKube, capability dynatracev1beta1.ActiveGateCapability, wantEnabled bool) {
+func SwitchCapability(instance *dynatracev1beta1.DynaKube, capability dynatracev1beta1.ActiveGateCapability, wantEnabled bool) {
 	hasEnabled := instance.IsActiveGateMode(capability.DisplayName)
 	capabilities := &instance.Spec.ActiveGate.Capabilities
 
@@ -13,11 +11,11 @@ func DoTestSetCapability(instance *dynatracev1beta1.DynaKube, capability dynatra
 	}
 
 	if !wantEnabled && hasEnabled {
-		*capabilities = testRemoveCapability(*capabilities, capability.DisplayName)
+		*capabilities = removeCapability(*capabilities, capability.DisplayName)
 	}
 }
 
-func testRemoveCapability(capabilities []DtCapability, removeMe DtCapability) []DtCapability {
+func removeCapability(capabilities []dynatracev1beta1.CapabilityDisplayName, removeMe dynatracev1beta1.CapabilityDisplayName) []dynatracev1beta1.CapabilityDisplayName {
 	for i, capability := range capabilities {
 		if capability == removeMe {
 			return append(capabilities[:i], capabilities[i+1:]...)
